+++ src/lib.rs
@@ -23,9 +23,9 @@
 //! ## Example
 //! ```
 //! use miniconf::{Miniconf, MiniconfAtomic};
-//! use serde::Deserialize;
+//! use serde::{Serialize, Deserialize};
 //!
-//! #[derive(Deserialize, MiniconfAtomic, Default)]
+//! #[derive(Deserialize, Serialize, MiniconfAtomic, Default)]
 //! struct Coefficients {
 //!     forward: f32,
 //!     backward: f32,
@@ -42,13 +42,13 @@
 //! let mut settings = Settings::default();
 //!
 //! // Update sample rate.
-//! miniconf::update(&mut settings, "sample_rate", b"350").unwrap();
+//! settings.set("sample_rate", b"350").unwrap();
 //!
 //! // Update filter coefficients.
-//! miniconf::update(&mut settings, "filter", b"{\"forward\": 35.6, \"backward\": 0.0}").unwrap();
+//! settings.set("filter", b"{\"forward\": 35.6, \"backward\": 0.0}").unwrap();
 //!
 //! // Update channel gain for channel 0.
-//! miniconf::update(&mut settings, "channel_gain/0", b"15").unwrap();
+//! settings.set("channel_gain/0", b"15").unwrap();
 //! ```
 //!
 //! ## Limitations
@@ -60,6 +60,8 @@
 #[cfg(feature = "mqtt-client")]
 mod mqtt_client;
 
+pub mod iter;
+
 #[cfg(feature = "mqtt-client")]
 pub use mqtt_client::MqttClient;
 
@@ -70,13 +72,18 @@ pub use minimq;
 pub use minimq::embedded_time;
 
 #[doc(hidden)]
-pub use serde::de::{Deserialize, DeserializeOwned};
+pub use serde::{
+    de::{Deserialize, DeserializeOwned},
+    ser::Serialize,
+};
 
 pub use serde_json_core;
 
 pub use derive_miniconf::{Miniconf, MiniconfAtomic};
 
 pub use heapless;
+
+use core::fmt::Write;
 
 /// Errors that occur during settings configuration
 #[derive(Debug, PartialEq)]
@@ -107,6 +114,11 @@ pub enum Error {
     /// Check that the serialized data is valid JSON and of the correct type.
     Deserialization(serde_json_core::de::Error),
 
+    /// The value provided could not be serialized.
+    ///
+    /// Check that the buffer had sufficient space.
+    SerializationFailed,
+
     /// When indexing into an array, the index provided was out of bounds.
     ///
     /// Check array indices to ensure that bounds for all paths are respected.
@@ -122,6 +134,7 @@ impl From<Error> for u8 {
             Error::AtomicUpdateRequired => 4,
             Error::Deserialization(_) => 5,
             Error::BadIndex => 6,
+            Error::SerializationFailed => 7,
         }
     }
 }
@@ -133,35 +146,49 @@ impl From<serde_json_core::de::Error> for Error {
 }
 
 pub trait Miniconf {
+    /// Update settings directly from a string path and data.
+    ///
+    /// # Args
+    /// * `path` - The path to update within `settings`.
+    /// * `data` - The serialized data making up the contents of the configured value.
+    ///
+    /// # Returns
+    /// The result of the configuration operation.
+    fn set(&mut self, path: &str, data: &[u8]) -> Result<(), Error> {
+        self.string_set(path.split('/').peekable(), data)
+    }
+
+    fn get(&self, path: &str, data: &mut [u8]) -> Result<usize, Error> {
+        self.string_get(path.split('/').peekable(), data)
+    }
+
+    fn into_iter<'a, const TS: usize>(
+        &'a self,
+        state: &'a mut [usize],
+    ) -> iter::MiniconfIter<'a, Self, TS> {
+        iter::MiniconfIter {
+            settings: self,
+            state,
+        }
+    }
+
     fn string_set(
         &mut self,
         topic_parts: core::iter::Peekable<core::str::Split<char>>,
         value: &[u8],
     ) -> Result<(), Error>;
 
-    fn recursive_iter<const TS: usize, const VS: usize>(
+    fn string_get(
+        &self,
+        topic_parts: core::iter::Peekable<core::str::Split<char>>,
+        value: &mut [u8],
+    ) -> Result<usize, Error>;
+
+    fn recursive_iter<const TS: usize>(
         &self,
         index: &mut [usize],
-        _topic: &mut heapless::String<TS>,
-        value: &mut heapless::String<VS>,
+        topic: &mut heapless::String<TS>,
     ) -> Option<()>;
-}
-
-/// Convenience function to update settings directly from a string path and data.
-///
-/// # Note
-/// When using prefixes on the path, it is often simpler to call
-/// `Settings::string_set(path.peekable(), data)` directly.
-///
-/// # Args
-/// * `settings` - The settings to update
-/// * `path` - The path to update within `settings`.
-/// * `data` - The serialized data making up the contents of the configured value.
-///
-/// # Returns
-/// The result of the configuration operation.
-pub fn update<T: Miniconf>(settings: &mut T, path: &str, data: &[u8]) -> Result<(), Error> {
-    settings.string_set(path.split('/').peekable(), data)
 }
 
 macro_rules! impl_single {
@@ -179,17 +206,25 @@ macro_rules! impl_single {
                 Ok(())
             }
 
+            fn string_get(
+                &self,
+                mut topic_parts: core::iter::Peekable<core::str::Split<char>>,
+                value: &mut [u8],
+            ) -> Result<usize, Error> {
+                if topic_parts.peek().is_some() {
+                    return Err(Error::PathTooLong);
+                }
+
+                serde_json_core::to_slice(self, value).map_err(|_| Error::SerializationFailed)
+            }
+
             // This implementation is the base case for primitives where it will
             // yield once for self, then return None on subsequent calls.
-            fn recursive_iter<const TS: usize, const VS: usize>(
+            fn recursive_iter<const TS: usize>(
                 &self,
                 index: &mut [usize],
                 _topic: &mut heapless::String<TS>,
-                value: &mut heapless::String<VS>,
-            ) -> Option<()>
-            where
-                Self: serde::Serialize,
-            {
+            ) -> Option<()> {
                 if index.len() == 0 {
                     // I don't expect this to happen...
                     panic!("index stack too small");
@@ -200,12 +235,10 @@ macro_rules! impl_single {
                 index[0] += 1;
                 index[1..].iter_mut().for_each(|x| *x = 0);
 
-                match i {
-                    0 => {
-                        *value = serde_json_core::to_string(&self).unwrap();
-                        Some(())
-                    }
-                    _ => None,
+                if i == 0 {
+                    Some(())
+                } else {
+                    None
                 }
             }
         }
@@ -237,13 +270,59 @@ impl<T: Miniconf, const N: usize> Miniconf for [T; N] {
         Ok(())
     }
 
-    fn recursive_iter<const TS: usize, const VS: usize>(
+    fn string_get(
         &self,
-        _index: &mut [usize],
-        _topic: &mut heapless::String<TS>,
-        _value: &mut heapless::String<VS>,
+        mut topic_parts: core::iter::Peekable<core::str::Split<char>>,
+        value: &mut [u8],
+    ) -> Result<usize, Error> {
+        let next = topic_parts.next();
+        if next.is_none() {
+            return Err(Error::PathTooShort);
+        }
+
+        // Parse what should be the index value
+        let i: usize = serde_json_core::from_str(next.unwrap())
+            .or(Err(Error::BadIndex))?
+            .0;
+
+        if i >= self.len() {
+            return Err(Error::BadIndex);
+        }
+
+        self[i].string_get(topic_parts, value)
+    }
+
+    fn recursive_iter<const TS: usize>(
+        &self,
+        index: &mut [usize],
+        topic: &mut heapless::String<TS>,
     ) -> Option<()> {
-        unimplemented!("iteration not implemented for arrays yet")
+        while index[0] < N {
+            // Add the array index to the topic name.
+            if write!(topic, "/{}", index[0]).is_err() {
+                return None;
+            }
+
+            if self[index[0]]
+                .recursive_iter(&mut index[1..], topic)
+                .is_some()
+            {
+                return Some(());
+            }
+
+            // Strip off the previously prepended index, since we completed that element and need
+            // to instead check the next one.
+            while let Some(character) = topic.pop() {
+                if character == '/' {
+                    break;
+                }
+            }
+
+            index[0] += 1;
+            index[1..].iter_mut().for_each(|x| *x = 0);
+        }
+
+        None
     }
 }