+++ tests/enums.rs
@@ -1,15 +1,15 @@
 use miniconf::Miniconf;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
 #[test]
 fn simple_enum() {
-    #[derive(Miniconf, Debug, Deserialize, PartialEq)]
+    #[derive(Miniconf, Debug, Deserialize, Serialize, PartialEq)]
     enum Variant {
         A,
         B,
     }
 
-    #[derive(Miniconf, Debug, Deserialize)]
+    #[derive(Miniconf, Debug, Deserialize, Serialize)]
     struct S {
         v: Variant,
     }
@@ -25,7 +25,7 @@ fn simple_enum() {
 
 #[test]
 fn invalid_enum() {
-    #[derive(Miniconf, Debug, Deserialize, PartialEq)]
+    #[derive(Miniconf, Debug, Serialize, Deserialize, PartialEq)]
     enum Variant {
         A,
         B,